refactor: creates new files for structs and helpers
